requirement.txt
python3-flask=2.2.2
python3-flask-socketio=5.3.2
python3-flask-cors=3.0.10
python3-psutil=5.9.4
python3-picamera2=0.3.21
python3-libcamera
python3-eventlet=0.33.1
python3-twisted=22.4.0
python3-cryptography=38.0.4
python3-opencv
python3-serial
xdg-utils



pip3 install (requirement.txt)
sudo apt install -y xdg-utils
apt install unzip
mkdir app
cd app
git clone https://github.com/BFMC-2025-COSS/Brain.git
cd Brain
git checkout 9-make-brain-work-with-cv2-not-picamera2


cd src/dashboard/frontend/
curl -fsSL https://fnm.vercel.app/install | bash
export PATH="/root/.local/share/fnm/node-versions/v22.13.0/installation/bin:$PATH"

fnm install --lts
npm install -g @angular/cli@17
npm install

cd ../../..
